Move defaultFilters back
